adress review comments
