+++ pkg/deployment/depresource/stop.go
@@ -21,7 +21,6 @@ import (
 	"github.com/elastic/cloud-sdk-go/pkg/api"
 	"github.com/elastic/cloud-sdk-go/pkg/client/deployments"
 	"github.com/elastic/cloud-sdk-go/pkg/models"
-	"github.com/elastic/cloud-sdk-go/pkg/util/slice"
 	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 
@@ -49,12 +48,12 @@ func (params StopParams) Validate() error {
 		merr = multierror.Append(merr, deputil.NewInvalidDeploymentIDError(params.DeploymentID))
 	}
 
-	if !slice.HasString(util.ValidTypes, params.Type) {
-		merr = multierror.Append(merr, deputil.NewInvalidResourceTypeError(params.Type))
+	if params.Type == "" {
+		merr = multierror.Append(merr, errors.New("deployment resource type cannot be empty"))
 	}
 
 	if params.RefID == "" {
-		merr = multierror.Append(merr, errors.New("deployment stop: a ref_id must be provided"))
+		merr = multierror.Append(merr, errors.New("a ref_id must be provided"))
 	}
 
 	return merr.ErrorOrNil()
@@ -108,48 +107,6 @@ func StopInstances(params StopInstancesParams) (models.DeploymentResourceCommand
 
 	res, err := params.V1API.Deployments.StopDeploymentResourceInstances(
 		deployments.NewStopDeploymentResourceInstancesParams().
-			WithDeploymentID(params.DeploymentID).
-			WithResourceKind(params.Type).
-			WithIgnoreMissing(params.IgnoreMissing).
-			WithInstanceIds(params.InstanceIDs).
-			WithRefID(params.RefID),
-		params.AuthWriter,
-	)
-	if err != nil {
-		return nil, api.UnwrapError(err)
-	}
-
-	return res.Payload, nil
-}
-
-// StopMaintenanceMode stops maintenance mode of all instances belonging to a deployment resource type.
-func StopMaintenanceMode(params StopParams) (models.DeploymentResourceCommandResponse, error) {
-	if err := params.Validate(); err != nil {
-		return nil, err
-	}
-
-	res, err := params.V1API.Deployments.StopDeploymentResourceInstancesAllMaintenanceMode(
-		deployments.NewStopDeploymentResourceInstancesAllMaintenanceModeParams().
-			WithDeploymentID(params.DeploymentID).
-			WithResourceKind(params.Type).
-			WithRefID(params.RefID),
-		params.AuthWriter,
-	)
-	if err != nil {
-		return nil, api.UnwrapError(err)
-	}
-
-	return res.Payload, nil
-}
-
-// StopInstancesMaintenanceMode stops maintenance mode of defined instances belonging to a deployment resource.
-func StopInstancesMaintenanceMode(params StopInstancesParams) (models.DeploymentResourceCommandResponse, error) {
-	if err := params.Validate(); err != nil {
-		return nil, err
-	}
-
-	res, err := params.V1API.Deployments.StopDeploymentResourceMaintenanceMode(
-		deployments.NewStopDeploymentResourceMaintenanceModeParams().
 			WithDeploymentID(params.DeploymentID).
 			WithResourceKind(params.Type).
 			WithIgnoreMissing(params.IgnoreMissing).
